pyproject.toml
[project]
name = "quant-strategy-app"
version = "0.1.0"
description = "Template for strategy-agnostic quant research, backtesting, and live execution."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic",
  "pandas",
  "pydantic-settings",
  "pyyaml",
]

[project.scripts]
qsa = "qsa.cli:main"

[dependency-groups]
dev = [
  "mypy",
  "pandas-stubs",
  "pytest",
  "pytest-cov",
  "ruff",
  "types-PyYAML",
]

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["qsa*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
strict_equality = true
